requirements.txt
kaleido==0.2.1
lempel_ziv_complexity==0.2.2
numpy==1.25.1
pandas==2.0.3
plotly==5.15.0
quine_mccluskey==0.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
wandb==0.15.7
argparse==1.4.0
pyYAML==6.0.1
tqdm==4.65.0
matplotlib==3.7.1
scipy==1.11.1
scikit-learn==1.3.0
einops==0.6.1
hydra-core==1.3.2
omegaconf==2.3.0
pytest==7.3.1
